pyproject.toml
[project]
name = "inbox"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12,<3.14"
dependencies = [
    "fastapi>=0.115.0",
    "google-api-python-client>=2.194.0",
    "google-auth-oauthlib>=1.3.1",
    "httpx>=0.28.0",
    "loguru>=0.7.3",
    "mlx-lm>=0.21.0",
    "mlx-whisper>=0.4.0",
    "numpy>=1.26.0",
    "outlines>=0.1.0",
    "pydantic>=2.12.5",
    "pyobjc-framework-Quartz>=10.0",
    "python-multipart>=0.0.24",
    "rich>=14.3.3",
    "sounddevice>=0.5.0",
    "textual>=8.2.3",
    "uvicorn>=0.34.0",
]

[dependency-groups]
dev = [
    "bandit>=1.9.4",
    "hypothesis>=6.151.12",
    "pre-commit>=4.5.1",
    "pyright>=1.1.408",
    "pytest>=9.0.3",
    "pytest-cov>=7.1.0",
    "python-dotenv>=1.2.2",
    "ruff>=0.15.10",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "basic"
reportMissingImports = true
reportUnusedImports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=. --cov-report=term-missing"

[tool.bandit]
exclude_dirs = [".venv", "tests"]
skips = ["B101", "B110", "B112", "B404", "B603", "B607"]
